Desktop/fyp/xssniper_supabase/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
starlette
httpx
itsdangerous
supabase>=2.0.0
python-dotenv
